Move error classes into their own header
